+++ CHANGES.md
@@ -1,4 +1,7 @@
 
+  - added `mul` and `div`
+  - added `print`
+  - fixed code and doc of `map` and `mapi` (now fails if `f` returns 0)
   - improved documentation: applicative data type / complexity
 
 # 1.0.0